pyproject.toml
[project]
name = "prod"
version = "0.1.0"
description = "prod client"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "flet==0.25.2",
  "opencv-python>=4.10.0",
  "pillow>=10.4.0",
  "requests>=2.32.3",
  #"sqlite3>=2.6.0"
  #"zxing_cpp>=2.3.0"
  #"pyzbar>=0.1.9",#"libzbar0t64>=0.23"
]

[tool.flet]
org = "com.github.usermicrodevices"
# shown in window titles and about app dialogs on desktop.
product = "prod-client"
# company name to display in about app dialogs
company = "prod"
# copyright text to display in about app dialogs
copyright = "github.com/usermicrodevices/prod"
permissions = ["camera"]

[tool.flet.app]
path = "src"

[tool.flet.android.permission]
"android.permission.CAMERA" = true

[tool.flet.android.feature]
"android.hardware.camera" = false
